Fix citystations leaks

+++ helpers/recalloc.c
@@ -6,7 +6,7 @@ void * recalloc(void * vec, size_t len, size_t new_len){
 
 	if(new == NULL) return NULL;
 
-	for(int i = len; i < new_len; i++){
+	for(size_t i = len; i < new_len; i++){
 		((char *)new)[i] = 0;
 	}
